changed id type from integer to long

+++ ticket-service/src/main/java/com/epam/training/ticketservice/repository/ScreeningRepository.java
@@ -9,7 +9,7 @@ import org.springframework.data.jpa.repository.Query;
 import java.time.LocalDateTime;
 import java.util.Optional;
 
-public interface ScreeningRepository extends JpaRepository<Screening, Integer> {
+public interface ScreeningRepository extends JpaRepository<Screening, Long> {
     @Query(value = "FROM Screening\nWHERE (MOVIE_NAME = ?1 AND ROOM_NAME = ?2 AND SCREENING_START = ?3)")
     Optional<Screening> findOne(Movie movie, Room room, LocalDateTime start);
 }
